Closes #2 Closes #3 Closes #5
Now detects and migrates some tool packages to RC2.
Now upgrades framework in project.json, by renaming dnx4YZ to net4YZ, as well as adding the appropriate netcoreapp1.0 or netstandard1.5 TFM, based on whether the project is a Library or an Application (i.e whether it has an entry point).
Now ensues buildOptions are present and sets preserveCompilationContext to true by default.